refactor(gateway): share tracked-run ownership checks
