+++ src/gateway/server-plugins.ts
@@ -10,7 +10,11 @@ import { setActivePluginRegistry } from "../plugins/runtime.js";
 import { getPluginRuntimeGatewayRequestScope } from "../plugins/runtime/gateway-request-scope.js";
 import type { PluginRuntime } from "../plugins/runtime/types.js";
 import { resolveGlobalSingleton } from "../shared/global-singleton.js";
-import { abortTrackedRunById, normalizeOptionalTrackedText } from "./chat-abort.js";
+import {
+  abortTrackedRunById,
+  canRequesterAccessTrackedRun,
+  normalizeOptionalTrackedText,
+} from "./chat-abort.js";
 import { ADMIN_SCOPE, WRITE_SCOPE } from "./method-scopes.js";
 import { GATEWAY_CLIENT_IDS, GATEWAY_CLIENT_MODES } from "./protocol/client-info.js";
 import type { ErrorShape } from "./protocol/index.js";
@@ -243,29 +247,6 @@ function hasAdminScope(client: GatewayRequestOptions["client"] | null | undefine
   return scopes.includes(ADMIN_SCOPE);
 }
 
-function canClientAbortTrackedAgentRun(params: {
-  client: GatewayRequestOptions["client"] | undefined;
-  active: { ownerConnId?: string; ownerDeviceId?: string };
-}): boolean {
-  if (hasAdminScope(params.client)) {
-    return true;
-  }
-  const ownerDeviceId = normalizeOptionalTrackedText(params.active.ownerDeviceId);
-  const ownerConnId = normalizeOptionalTrackedText(params.active.ownerConnId);
-  if (!ownerDeviceId && !ownerConnId) {
-    return true;
-  }
-  const requesterDeviceId = normalizeOptionalTrackedText(params.client?.connect?.device?.id);
-  if (ownerDeviceId && requesterDeviceId && ownerDeviceId === requesterDeviceId) {
-    return true;
-  }
-  const requesterConnId = normalizeOptionalTrackedText(params.client?.connId);
-  if (ownerConnId && requesterConnId && ownerConnId === requesterConnId) {
-    return true;
-  }
-  return false;
-}
-
 function canClientUseModelOverride(client: GatewayRequestOptions["client"]): boolean {
   return hasAdminScope(client) || client?.internal?.allowModelOverride === true;
 }
@@ -330,7 +311,13 @@ export function createGatewayAgentAbort(): PluginRuntime["agent"]["abort"] {
     if (!active || active.kind !== "agent") {
       return { aborted: false };
     }
-    if (!canClientAbortTrackedAgentRun({ client: scope?.client, active })) {
+    if (
+      !canRequesterAccessTrackedRun(active, {
+        connId: normalizeOptionalTrackedText(scope?.client?.connId),
+        deviceId: normalizeOptionalTrackedText(scope?.client?.connect?.device?.id),
+        isAdmin: hasAdminScope(scope?.client),
+      })
+    ) {
       return { aborted: false };
     }
     return abortTrackedRunById(ctx, {
